fix issue where input border is cut or half hidden (#143)
A few browsers have as its behaviour to round its size when passing relative sizes (e.g. 100%)
so if the iframe itself fits a container or have a width with decimal case (e.g. 350.55px),
the iframe inner content will have a width of 351px, where in our case it endup cutting
the border at right.

This fix prevents that from happening by adding taking 1px of the input and adding as padding
so it have 1px of empty space